feat: allow route rules, specifying in/out interfaces and prepend
For service/application/interface rules it's now possible to specify an
input and output interface.
For interface the output interface is part of the key, something like
eth1:eth2.
It's also possible to define if the rule is actually a routing rule to
have ufw do a `ufw route allow|deny|limit`.
Finally it's possible to set prepend on a rule to put them at the top of
the chain, by default deny rules are prepended.